Move changing answers specs to their own file
Some more re-organising of the feature specs as having them all in the
same file was getting unwieldy.
